refactor(menu): reorganize context imports and create MenuContext component
- Updated import paths for MenuContext in MenuContent and MenuRoot components to reflect the new file structure.
- Introduced MenuContext component to manage menu-related state and actions, enhancing component reusability.
